refactor(python): split _internal/numpy_utils.py into a package
Convert the flat numpy_utils.py into a numpy_utils/ package (sentinel.py,
type.py, pad.py, pyarrow_conversion.py) mirroring the internal SDK's
shared/utils/numpy_utils/ layout, per Hong's review feedback on PR #1543.
The public import path (michelangelo.lib._internal.numpy_utils) is
unchanged via __init__.py re-exports, so no caller updates are needed.
